all bullets are set to "despawn" when coordinates are lower than 0 or greater than map's size (irrelevant but...) added getters to mainscreen's raw map width & raw map height

+++ TFG_GAME/core/src/com/alexgs/tfg_game/elements/tools/Weapons.java
@@ -3,7 +3,7 @@ package com.alexgs.tfg_game.elements.tools;
 public class Weapons {
     private int persistenceMagSize;
     private boolean canSwitchFireMode;
-    private int roundsPerMinute;
+    private float roundsPerMinute;
     private int dmg;
     private float timeAllowedToExist;
 
@@ -33,7 +33,7 @@ public class Weapons {
         return canSwitchFireMode;
     }
 
-    public int getRoundsPerMinute() {
+    public float getRoundsPerMinute() {
         return roundsPerMinute;
     }
